+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Unless otherwise noted, all changes below pertains to the Scripting feature in J
 - Add `customDestination` to `Stop`, which exposes the raw custom destination value.
 - Add `TickableSoundInstance`, allowing sound parameters (e.g. volume/pitch) to be continuously adjusted throughout the playback, as well as the ability to stop the sound midway.
   - `MinecraftClient.stopSound()` has been removed, you should use a TickableSoundInstance instead and stop with `SoundManager`.
+- **PIDS Scripting**
+  - Add `PIDSBlockEntity.getTargetPlatformIds()` to obtain the detected/selected platform id for ETA monitoring.
 
 ## Changes
 - `ModelData` has been renamed back to `RawModel`